Got test05 working for nested loops.  The problem was the control loop getting too far ahead of the data loop.  Fixed by making the last branch at the end of the for_body basic block conditional on a PredOp.  In this case, for both loops, the PredOp was a preceding store.  This means the increment for the control feedback loop doesn't fire until after the body is complete.
To make this work I also modified GEP, load, store, and compute nodes
to fire for inactive enables so that they propegate all the way through
the loop body even on the last iteration.  The loads and stores don't
actually execute but they do propegate a predicated output down the chain.